refactor(upload): seperate components and improve directory structure
