+++ shared/design/src/components/upload/components/index.ts
@@ -1,0 +1,8 @@
+export { FileUploadClear } from "./clear"
+export { FileUploadDropzone } from "./dropzone"
+export { FileUploadItem } from "./item"
+export { FileUploadItemDelete } from "./itemDelete"
+export { FileUploadItemMetadata } from "./itemMetadata"
+export { FileUploadItemProgress } from "./itemProgress"
+export { FileUploadList } from "./list"
+export { FileUploadTrigger } from "./trigger"
